Rebase

+++ src/main/java/com/shoploc/shoploc/domain/store/StoreServiceImpl.java
@@ -3,14 +3,18 @@ package com.shoploc.shoploc.domain.store;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.io.File;
 import java.util.List;
 import java.util.Optional;
 
 @Service
 public class StoreServiceImpl implements StoreService {
 
-    @Autowired
     private StoreRepository storeRepository;
+
+    public StoreServiceImpl (StoreRepository storeRepository) {
+        this.storeRepository = storeRepository;
+    }
 
     @Override
     public List<Store> getAllStores() {
@@ -18,26 +22,31 @@ public class StoreServiceImpl implements StoreService {
     }
 
     @Override
-    public Optional<Store> getById(Long id) {
-        return storeRepository.findById(id);
+    public Store getById(Long id) {
+        Optional<Store> store = storeRepository.findById(id);
+        if (store.isPresent()) {
+            return storeRepository.findById(id).get();
+        }
+        else return null;
     }
 
     @Override
-    public Store createStore(String name, String address) {
+    public Store createStore(String name, String address, File image) {
         Store store = new Store();
         store.setName(name);
         store.setAddress(address);
-
+        store.setImage(image);
         return storeRepository.save(store);
     }
 
     @Override
-    public Store updateStore(Long id, String name, String address) {
+    public Store updateStore(Long id, Store store) {
         Store existingStore = storeRepository.findById(id).orElse(null);
 
         if (existingStore != null) {
-            existingStore.setName(name);
-            existingStore.setAddress(address);
+            existingStore.setName(store.getName());
+            existingStore.setAddress(store.getAddress());
+            existingStore.setImage(store.getImage());
             return storeRepository.save(existingStore);
         }
 
@@ -45,7 +54,11 @@ public class StoreServiceImpl implements StoreService {
     }
 
     @Override
-    public void deleteById(Long id) {
-        storeRepository.deleteById(id);
+    public boolean deleteById(Long id) {
+        if (storeRepository.findById(id).isPresent()) {
+            storeRepository.deleteById(id);
+            return true;
+        }
+        else return false;
     }
 }
